chore: move snippets to a plugin folder, fix imports
