+++ plugins/official/index.ts
@@ -1,0 +1,1 @@
+export { stackSnippetPlugin } from "./stack-snippets/src/stackSnippetPlugin";
