+++ plugins/official/stack-snippets/src/stackSnippetPlugin.ts
@@ -1,5 +1,5 @@
 /** This is an effort to turn the existing structure back into an external plugin while keeping the same functionality*/
-import type { EditorPlugin } from "../../../src";
+import type { EditorPlugin } from "../../../../src";
 import {
     stackSnippetMarkdownParser,
     stackSnippetMarkdownSerializer,
@@ -10,51 +10,11 @@ import { Node as ProseMirrorNode } from "prosemirror-model";
 import { EditorView } from "prosemirror-view";
 import { StackSnippetView } from "./snippet-view";
 import { StackSnippetOptions } from "./common";
-import { error, log } from "../../../src/shared/logger";
 import { stackSnippetPasteHandler } from "./paste-handler";
-import { makeMenuButton } from "../../../src/shared/menu";
-import { _t } from "../../../src/shared/localization";
-import { getShortcut } from "../../../src/shared/utils";
+import { makeMenuButton } from "../../../../src/shared/menu";
+import { _t } from "../../../../src/shared/localization";
+import { getShortcut } from "../../../../src/shared/utils";
 import { openSnippetModal } from "./commands";
-
-export const demoOpts: StackSnippetOptions = {
-    renderer: (meta, js, css, html) => {
-        const data = {
-            js: js,
-            css: css,
-            html: html,
-            console: meta.console,
-            babel: meta.babel,
-            babelPresetReact: meta.babelPresetReact,
-            babelPresetTS: meta.babelPresetTS,
-        };
-        return fetch("/snippets/js", {
-            method: "POST",
-            body: new URLSearchParams(data),
-        })
-            .then((res) => res.text())
-            .then((html) => {
-                const parser = new DOMParser();
-                const doc = parser.parseFromString(html, "text/html");
-                return doc;
-            })
-            .catch((err) => {
-                error("test harness - snippet render", err);
-                const div = document.createElement("div");
-                const freeRealEstate = document.createElement("img");
-                freeRealEstate.src =
-                    "https://i.kym-cdn.com/entries/icons/original/000/021/311/free.jpg";
-                div.appendChild(freeRealEstate);
-                return div;
-            });
-    },
-    openSnippetsModal: (meta, js, css, html) => {
-        log("test harness - open modal event", `meta\n${JSON.stringify(meta)}`);
-        log("test harness - open modal event", `js\n${JSON.stringify(js)}`);
-        log("test harness - open modal event", `css\n${JSON.stringify(css)}`);
-        log("test harness - open modal event", `html\n${JSON.stringify(html)}`);
-    },
-};
 
 /**
  * Build the StackSnippet plugin using hoisted options that can be specified at runtime
