+++ plugins/official/stack-snippets/test/menu.test.ts
@@ -1,6 +1,6 @@
-import { createMenuEntries, MenuBlock } from "../../../src/shared/menu";
-import { EditorType } from "../../../src";
-import { testRichTextSchema } from "../../../test/rich-text/test-helpers";
+import { createMenuEntries, MenuBlock } from "../../../../src/shared/menu";
+import { EditorType } from "../../../../src";
+import { testRichTextSchema } from "../../../../test/rich-text/test-helpers";
 import { buildSnippetMenuEntries } from "./stack-snippet-helpers";
 
 function getEntryByKey(blocks: MenuBlock[], key: string) {
